Move line style to its own module
